cleaned up store
