+++ client/components/CartItem.js
@@ -1,7 +1,7 @@
 import React, {useState} from 'react'
 import {Link} from 'react-router-dom'
 import { useDispatch } from 'react-redux'
-import {thunkUpdateOrder, thunkRemoveFromOrder} from '../store'
+import {thunkUpdateOrder, thunkRemoveFromOrder} from '../store/thunks'
 
 export default function CartItem({item}) {
   const [quantity, setQuantity] = useState(item.quantity)
