+++ client/components/ProductCard.js
@@ -1,7 +1,7 @@
 import React from 'react'
 import {Link} from 'react-router-dom'
 import {connect} from 'react-redux'
-import {thunkCreateOrder, thunkAddToOrder, thunkLoadMyCart} from '../store'
+import {thunkCreateOrder, thunkAddToOrder, thunkLoadMyCart} from '../store/thunks'
 import axios from 'axios'
