+++ client/store/constants.js
@@ -1,0 +1,35 @@
+const LOAD_PRODUCTS = 'LOAD_PRODUCTS'
+const CREATE_PRODUCT = 'CREATE_PRODUCT'
+const UPDATE_PRODUCT = 'UPDATE_PRODUCT'
+const REMOVE_PRODUCT = 'REMOVE_PRODUCT'
+
+const LOAD_CATEGORIES = 'LOAD_CATEGORIES'
+
+const LOAD_ORDERS = 'LOAD_ORDERS'
+const LOAD_CART = 'LOAD_CART'
+const UPDATE_ORDER = 'UPDATE_ORDER'
+
+const CREATE_ORDER = 'CREATE_ORDER'
+const ADD_TO_ORDER = 'ADD_TO_ORDER'
+const REMOVE_FROM_ORDER = 'REMOVE_FROM_ORDER'
+const UPDATE_ORDER_STATUS = 'UPDATE_ORDER_STATUS'
+
+const LOAD_USERS = 'LOAD_USERS'
+const REMOVE_USER = 'REMOVE_USER'
+
+export {
+    LOAD_PRODUCTS,
+    CREATE_PRODUCT,
+    UPDATE_PRODUCT,
+    REMOVE_PRODUCT,
+    LOAD_CATEGORIES,
+    LOAD_ORDERS,
+    LOAD_CART,
+    UPDATE_ORDER,
+    CREATE_ORDER,
+    ADD_TO_ORDER,
+    REMOVE_FROM_ORDER,
+    UPDATE_ORDER_STATUS,
+    LOAD_USERS,
+    REMOVE_USER
+}
