+++ client/store/index.js
@@ -3,128 +3,122 @@ import {createLogger} from 'redux-logger'
 import thunkMiddleware from 'redux-thunk'
 import {composeWithDevTools} from 'redux-devtools-extension'
 import user from './user'
-import axios from 'axios'
 
-//Constants
-const LOAD_PRODUCTS = 'LOAD_PRODUCTS'
-const CREATE_PRODUCT = 'CREATE_PRODUCT'
-const UPDATE_PRODUCT = 'UPDATE_PRODUCT'
-const REMOVE_PRODUCT = 'REMOVE_PRODUCT'
+import {
+  LOAD_PRODUCTS,
+  CREATE_PRODUCT,
+  UPDATE_PRODUCT,
+  REMOVE_PRODUCT,
+  LOAD_CATEGORIES,
+  LOAD_ORDERS,
+  LOAD_CART,
+  UPDATE_ORDER,
+  CREATE_ORDER,
+  ADD_TO_ORDER,
+  REMOVE_FROM_ORDER,
+  UPDATE_ORDER_STATUS,
+  LOAD_USERS,
+  REMOVE_USER
+} from './constants'
 
-const LOAD_CATEGORIES = 'LOAD_CATEGORIES'
+// import {
+//   thunkLoadProducts,
+//   thunkCreateProduct,
+//   thunkRemoveProduct,
+//   thunkUpdateProduct,
+//   thunkLoadCategories,
+//   thunkLoadAllOrders,
+//   thunkLoadMyOrders,   
+//   thunkUpdateOrderStatus,
+//   thunkLoadUsers,
+//   thunkRemoveUser,
+//   thunkLoadMyCart,
+//   thunkCreateOrder,  
+//   thunkAddToOrder,
+//   thunkRemoveFromOrder,
+//   thunkUpdateOrder
+// } from './thunks'
 
-const LOAD_ORDERS = 'LOAD_ORDERS'
-const LOAD_CART = 'LOAD_CART'
-const UPDATE_ORDER = 'UPDATE_ORDER'
-
-const CREATE_ORDER = 'CREATE_ORDER'
-const ADD_TO_ORDER = 'ADD_TO_ORDER'
-const REMOVE_FROM_ORDER = 'REMOVE_FROM_ORDER'
-const UPDATE_ORDER_STATUS = 'UPDATE_ORDER_STATUS'
-
-const LOAD_USERS = 'LOAD_USERS'
-const REMOVE_USER = 'REMOVE_USER'
-
-
-//Action Creators
-const actionLoadProducts = products => ({type: LOAD_PRODUCTS, products})
-const actionCreateProduct = product => ({type: CREATE_PRODUCT, product})
-const actionUpdateProduct = product => ({type: UPDATE_PRODUCT, product})
-const actionRemoveProduct = id => ({type: REMOVE_PRODUCT, id})
-
-const actionLoadCategories = categories => ({type: LOAD_CATEGORIES, categories})
-
-const actionLoadOrders = orders => ({type: LOAD_ORDERS, orders})
-const actionLoadCart = cart => ({type: LOAD_CART, cart})
-const actionCreateOrder = order => ({type: CREATE_ORDER, order})
-const actionAddToOrder = order => ({type: ADD_TO_ORDER, order})
-
-const actionRemoveFromOrder = id => ({type: REMOVE_FROM_ORDER, id})
-const actionUpdateOrder = order => ({type: UPDATE_ORDER, order})
-const actionUpdateOrderStatus = order => ({type: UPDATE_ORDER_STATUS, order})
-
-const actionLoadUsers = users => ({type: LOAD_USERS, users})
-const actionRemoveUser = id => ({type: REMOVE_USER, id})
 
 //Thunks
-const thunkLoadProducts = () => async dispatch => {
-  const products = (await axios.get('/api/products')).data
-  return dispatch(actionLoadProducts(products))
-}
-const thunkCreateProduct = product => async dispatch => {
-  const newProduct = (await axios.post('/api/products', product)).data
-  dispatch(actionCreateProduct(newProduct))
-}
-const thunkRemoveProduct = id => async dispatch => {
-  await axios.delete(`/api/products/${id}`)
-  dispatch(actionRemoveProduct(id))
-}
-const thunkUpdateProduct = product => async dispatch => {
-  const currentProduct = (await axios.put(
-    `/api/products/${product.id}`,
-    product
-  )).data
-  dispatch(actionUpdateProduct(currentProduct))
-}
+// const thunkLoadProducts = () => async dispatch => {
+//   const products = (await axios.get('/api/products')).data
+//   return dispatch(actionLoadProducts(products))
+// }
+// const thunkCreateProduct = product => async dispatch => {
+//   const newProduct = (await axios.post('/api/products', product)).data
+//   dispatch(actionCreateProduct(newProduct))
+// }
+// const thunkRemoveProduct = id => async dispatch => {
+//   await axios.delete(`/api/products/${id}`)
+//   dispatch(actionRemoveProduct(id))
+// }
+// const thunkUpdateProduct = product => async dispatch => {
+//   const currentProduct = (await axios.put(
+//     `/api/products/${product.id}`,
+//     product
+//   )).data
+//   dispatch(actionUpdateProduct(currentProduct))
+// }
 
-const thunkLoadCategories = () => async dispatch => {
-  const categories = (await axios.get('/api/categories')).data
-  return dispatch(actionLoadCategories(categories))
-}
+// const thunkLoadCategories = () => async dispatch => {
+//   const categories = (await axios.get('/api/categories')).data
+//   return dispatch(actionLoadCategories(categories))
+// }
 
-const thunkLoadAllOrders = () => async dispatch => {
-  const orders = (await axios.get('/api/orders')).data
-  return dispatch(actionLoadOrders(orders))
-}
+// const thunkLoadAllOrders = () => async dispatch => {
+//   const orders = (await axios.get('/api/orders')).data
+//   return dispatch(actionLoadOrders(orders))
+// }
 
-const thunkLoadMyOrders = id => async dispatch => {
-  const orders = (await axios.get(`/api/orders/user/${id}`)).data
-  return dispatch(actionLoadOrders(orders))
-}
+// const thunkLoadMyOrders = id => async dispatch => {
+//   const orders = (await axios.get(`/api/orders/user/${id}`)).data
+//   return dispatch(actionLoadOrders(orders))
+// }
 
-const thunkUpdateOrderStatus = order => async dispatch => {
-  const currentOrder = (await axios.put(
-    `/api/orders/user/${order.id}`,
-    order
-  )).data
-  dispatch(actionUpdateOrderStatus(currentOrder))
-}
+// const thunkUpdateOrderStatus = order => async dispatch => {
+//   const currentOrder = (await axios.put(
+//     `/api/orders/user/${order.id}`,
+//     order
+//   )).data
+//   dispatch(actionUpdateOrderStatus(currentOrder))
+// }
 
-const thunkLoadUsers = () => {
-  return async(dispatch) => {
-    const users = (await axios.get('/api/users')).data
-    return dispatch(actionLoadUsers(users))
-  }
-}
-const thunkRemoveUser = id => async dispatch => {
-  await axios.delete(`/api/users/${id}`)
-  dispatch(actionRemoveUser(id))
-}
+// const thunkLoadUsers = () => {
+//   return async(dispatch) => {
+//     const users = (await axios.get('/api/users')).data
+//     return dispatch(actionLoadUsers(users))
+//   }
+// }
+// const thunkRemoveUser = id => async dispatch => {
+//   await axios.delete(`/api/users/${id}`)
+//   dispatch(actionRemoveUser(id))
+// }
 
-const thunkLoadMyCart = id => async dispatch => {
-  const cart = (await axios.get(`/api/orders/cart/${id}`)).data
-  return dispatch(actionLoadCart(cart))
-}
+// const thunkLoadMyCart = id => async dispatch => {
+//   const cart = (await axios.get(`/api/orders/cart/${id}`)).data
+//   return dispatch(actionLoadCart(cart))
+// }
 
-const thunkCreateOrder = (quantity, product, user) => async dispatch => {
-  const order = (await axios.post('/api/orders', {quantity, productId: product.id, userId: user.id})).data
-  return dispatch(actionCreateOrder(order))
-}
+// const thunkCreateOrder = (quantity, product, user) => async dispatch => {
+//   const order = (await axios.post('/api/orders', {quantity, productId: product.id, userId: user.id})).data
+//   return dispatch(actionCreateOrder(order))
+// }
 
-const thunkAddToOrder = (quantity, product, cartId) => async dispatch => {
-  const lineItem = (await axios.post('/api/lineItems', {quantity: quantity, productId: product.id, orderId: cartId})).data
-  return dispatch(actionAddToOrder(lineItem))
-}
+// const thunkAddToOrder = (quantity, product, cartId) => async dispatch => {
+//   const lineItem = (await axios.post('/api/lineItems', {quantity: quantity, productId: product.id, orderId: cartId})).data
+//   return dispatch(actionAddToOrder(lineItem))
+// }
 
-const thunkRemoveFromOrder = (id) => async dispatch => {
-  await axios.delete(`/api/lineItems/${id}`)
-  return dispatch(actionRemoveFromOrder(id))
-}
+// const thunkRemoveFromOrder = (id) => async dispatch => {
+//   await axios.delete(`/api/lineItems/${id}`)
+//   return dispatch(actionRemoveFromOrder(id))
+// }
 
-const thunkUpdateOrder = (quantity, id) => async dispatch => {
-  const lineItem = (await axios.put('/api/lineItems', {quantity, id})).data
-  return dispatch(actionUpdateOrder(lineItem))
-}
+// const thunkUpdateOrder = (quantity, id) => async dispatch => {
+//   const lineItem = (await axios.put('/api/lineItems', {quantity, id})).data
+//   return dispatch(actionUpdateOrder(lineItem))
+// }
 
 //Reducers
 const productReducer = (state = [], action) => {
@@ -230,20 +224,20 @@ const store = createStore(reducer, middleware)
 export default store
 export * from './user'
 
-export {
-  thunkLoadProducts,
-  thunkCreateProduct,
-  thunkUpdateProduct,
-  thunkRemoveProduct,
-  thunkLoadCategories,
-  thunkLoadAllOrders,
-  thunkLoadMyOrders,
-  thunkLoadUsers,
-  thunkRemoveUser,
-  thunkCreateOrder,
-  thunkLoadMyCart,
-  thunkAddToOrder,
-  thunkUpdateOrder,
-  thunkRemoveFromOrder,
-  thunkUpdateOrderStatus
-}
+// export {
+//   thunkLoadProducts,
+//   thunkCreateProduct,
+//   thunkUpdateProduct,
+//   thunkRemoveProduct,
+//   thunkLoadCategories,
+//   thunkLoadAllOrders,
+//   thunkLoadMyOrders,
+//   thunkLoadUsers,
+//   thunkRemoveUser,
+//   thunkCreateOrder,
+//   thunkLoadMyCart,
+//   thunkAddToOrder,
+//   thunkUpdateOrder,
+//   thunkRemoveFromOrder,
+//   thunkUpdateOrderStatus
+// }
