+++ public/style.css
@@ -91,7 +91,7 @@ input:focus {
   display: none;
   position: absolute;
   background-color: #495c6e;
-  width: 100px;
+  width: 150px;
   box-shadow: 0px 8px 16px 0px rgba(0, 0, 0, 0.5);
   z-index: 1;
   padding: 0.5rem;
